+++ template-parts/acf-blocks/postcard.php
@@ -1,0 +1,5 @@
+<?php
+
+$post_id = get_field( 'card_block_post_id' );
+
+lawyerist_get_post_card( $post_id, '', '', true );