+++ template-parts/loop-single.php
@@ -73,8 +73,6 @@ if ( have_posts() ) : while ( have_posts() ) : the_post();
 
   <?php
 
-  echo lawyerist_cta();
-
   lawyerist_get_related_resources();
 
 endwhile; endif; // Close the Loop.